+++ test/servers.js
@@ -1,7 +1,0 @@
-'use strict'
-
-let ramster = require('../index'),
-	config = require('./config'),
-	core = new ramster.Core(config)
-
-core.listen()